refactor: do not export utility types
