+++ packages/pure-parse/src/parsers/object/index.ts
@@ -1,0 +1,1 @@
+export * from './object'
