+++ packages/pure-parse/src/parsers/object/object.test.ts
@@ -1,27 +1,20 @@
 import { describe, expect, it, test } from 'vitest'
-import {
-  objectCompiled,
-  object,
-  Values,
-  Simplify,
-  OptionalKeys,
-  RequiredKeys,
-} from './object'
-import { isSuccess, Parser } from './types'
-import { Equals, OmitProperty, omitProperty } from '../internals'
-import { oneOf } from './oneOf'
+import { objectCompiled, object } from './object'
+import { isSuccess, Parser } from '../types'
+import { Equals, omitProperty } from '../../internals'
+import { oneOf } from '../oneOf'
 import {
   parseBoolean,
   parseNumber,
   parseString,
   parseUndefined,
-} from './primitives'
-import { Infer } from '../common'
-import { literal } from './literal'
-import { nullable, optional, undefineable } from './optional'
-import { objectMemo, objectCompiledMemo } from '../memoization'
-import { succeedWith, withDefault } from './defaults'
-import { parseUnknown } from './unknown'
+} from '../primitives'
+import { Infer } from '../../common'
+import { literal } from '../literal'
+import { nullable, optional, undefineable } from '../optional'
+import { objectMemo, objectCompiledMemo } from '../../memoization'
+import { succeedWith, withDefault } from '../defaults'
+import { parseUnknown } from '../unknown'
 
 const suites = [
   {
@@ -786,68 +779,5 @@ suites.forEach(({ name: suiteName, fn: object }) => {
     const n: Infer<typeof parseNode> = {
       data: 'a',
     }
-  })
-})
-
-describe('utility types', () => {
-  describe('RequiredKeys', () => {
-    it('extract required keys', () => {
-      const t0: Equals<
-        RequiredKeys<{
-          a: string
-          b: string | OmitProperty
-        }>,
-        'a'
-      > = true
-    })
-    it('handles required unknown', () => {
-      const t1: Equals<RequiredKeys<{ a: unknown }>, 'a'> = true
-      // @ts-expect-error -- the library cannot correctly infer optional unknowns. This is a compromise.
-      const t2: Equals<
-        RequiredKeys<{ a: unknown | OmitProperty }>,
-        never
-      > = true
-    })
-  })
-  describe('OptionalKeys', () => {
-    it('extract optional keys', () => {
-      const t0: Equals<
-        OptionalKeys<{
-          a: string
-          b: string | OmitProperty
-        }>,
-        'b'
-      > = true
-    })
-    it('handles optional unknown', () => {
-      const t1: Equals<
-        OptionalKeys<{
-          a: unknown
-        }>,
-        never
-      > = true
-      // @ts-expect-error -- the library cannot correctly infer optional unknowns. This is a compromise.
-      const t2: Equals<OptionalKeys<{ a: OmitProperty | unknown }>, 'a'> = true
-    })
-  })
-
-  describe('Values', () => {
-    it('handles empty objects', () => {
-      const t1: Equals<Values<{}>, never> = true
-    })
-    it('handles non-empty objects', () => {
-      const t1: Equals<Values<{ a: 1 }>, 1> = true
-      const t2: Equals<Values<{ a: 1; b: 2 }>, 1 | 2> = true
-    })
-    it('handles duplicate values', () => {
-      const t1: Equals<Values<{ a: 1; b: 1 }>, 1> = true
-    })
-  })
-  describe('Simplify', () => {
-    it('does not transform the type', () => {
-      const t1: Equals<Simplify<{}>, {}> = true
-      const t2: Equals<Simplify<{ a: string }>, { a: string }> = true
-      const t3: Equals<Simplify<string>, string> = true
-    })
   })
 })
